Had to modify to suite Pyslice.
How did this work before?  Somehow I was getting a '+ ' as a key, when
it needed to be '+'.  Tracked it down to the __get_iteration_and_command.
Reimplemented using .split() which automatically removes the spaces from
both sides of the values.

I suspect that Claudio can parse strings like '+variable_name...' which I
think should make an error.